imager/requirements.txt
boto3==1.12.36
botocore==1.15.36
certifi==2020.4.5.1
cffi==1.14.0
chardet==3.0.4
cryptography==2.9
docutils==0.15.2
idna==2.9
jmespath==0.9.5
ntlm-auth==1.4.0
pycparser==2.20
python-dateutil==2.8.1
pywinrm==0.4.1
requests==2.23.0
requests-ntlm==1.1.0
s3transfer==0.3.3
six==1.14.0
urllib3==1.25.8
xmltodict==0.12.0
